test(batched_pq): convert input ranges to vectors (#13)
Apparently we were going a bit overboard with all the `ranges::views` up to the point where some part of the program tried to take references of temporaries. Converting each test batch to a vector fixes these issues and also reflects what is passed to the batched PQ in reality.